# frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sync/manifest": {
      "get": {
        "operationId": "SyncController_getManifest",
        "parameters": [
          {
            "name": "X-Screen-ID",
            "in": "header",
            "description": "Screen identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/{screenId}/state": {
      "post": {
        "operationId": "SyncController_updateScreenState",
        "parameters": [
          {
            "name": "screenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScreenStateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/{screenId}/logs": {
      "post": {
        "operationId": "SyncController_recordLog",
        "parameters": [
          {
            "name": "screenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Sync"
        ]
      },
      "get": {
        "operationId": "SyncController_getLogs",
        "parameters": [
          {
            "name": "screenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/{screenId}/heartbeat": {
      "post": {
        "operationId": "SyncController_heartbeat",
        "parameters": [
          {
            "name": "screenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScreenHeartbeatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/{screenId}/check": {
      "get": {
        "operationId": "SyncController_checkManifestChanged",
        "parameters": [
          {
            "name": "screenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "currentRevision",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/{screenId}/fallback": {
      "get": {
        "operationId": "SyncController_getFallback",
        "parameters": [
          {
            "name": "screenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sync"
        ]
      }
    }
  },
  "info": {
    "title": "CMS API",
    "description": "API dla ekranów",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": ""
    }
  ],
  "components": {
    "schemas": {
...
      "UpdateScreenStateDto": {
        "type": "object",
        "properties": {

        }
      },
      "RecordLogDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Playback started",
            "description": "Treść komunikatu logu"
          },
          "level": {
            "type": "string",
            "enum": [
              "INFO",
              "ERROR",
              "PLAYBACK"
            ],
            "example": "INFO",
            "description": "Poziom logu"
          }
        },
        "required": [
          "message"
        ]
      },
      "UpdateScreenHeartbeatDto": {
        "type": "object",
        "properties": {

        }
      }
    }
  }
}